refactor: node traversal improvements
Separates node traversal into a separate function that constructs a very
basic table structure that can be used to determine what node to show in
cases where there are multiple per line.

Currently this uses the last node on the line, which is the reverse of
the changes made in #19, but will be more accurate.

This can be expanded upon by adding resolvers for spesific languages
etc.

The debugging feature now also uses this function to get more detailed
output.